requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
requests==2.31.0
transformers==4.37.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.6.0
google-generativeai==0.3.2
python-multipart==0.0.9
pydantic==2.6.1 